[solver] Move DirectInferenceAgent and TemplateManager headers to inference lib
