+++ problem-solver/cxx/inference-module/lib/include/inference/direct_inference_agent.hpp
