+++ problem-solver/cxx/inference-module/lib/include/inference/template_manager_abstract.hpp
