+++ problem-solver/cxx/inference-module/lib/src/manager/template-manager/TemplateManagerFixedArguments.hpp
@@ -6,11 +6,11 @@
 
 #pragma once
 
-#include "TemplateManagerAbstract.hpp"
+#include "inference/template_manager_abstract.hpp"
 
 #include <vector>
 
-#include "sc-memory/sc_memory.hpp"
+#include <sc-memory/sc_memory.hpp>
 
 namespace inference
 {
