+++ problem-solver/cxx/inference-module/test/units/TestLogicModuleArguments.cpp
@@ -4,9 +4,8 @@
  * (See accompanying file COPYING.MIT or copy at http://opensource.org/licenses/MIT)
  */
 
-#include "agent/DirectInferenceAgent.hpp"
-
-#include "inference/inference_keynodes.hpp"
+#include <inference/direct_inference_agent.hpp>
+#include <inference/inference_keynodes.hpp>
 
 #include <sc-memory/test/sc_test.hpp>
 #include <sc-builder/scs_loader.hpp>
